update docs
This commit does two things. Firstly the correct file path to transcompiled js file is /dist/index.js not /dist/src/index.js. Secondly for the example to run npm test needs to be called first in order to build into the correct directory. Although this is somewhat non-standard with the rest of helia-examples following some npm install then npm start as the convention. However I did not want to disrupt the rest of the docs.